requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.36
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.9
pytest==8.3.3
ruff==0.6.8
argon2-cffi==21.3.0
tabulate==0.9.0